+++ NEWS.md
@@ -1,4 +1,4 @@
-# teal.osprey 0.1.8.9000
+# teal.osprey 0.1.9
 * Added new `teal` module `tm_g_ae_sub` for an AE by subgroups plot.
 * Improved reactivity dependencies across modules.